Use format_exponential_notation from Console in Latex
